# Conflicts:
#	src/mat/dao/clause/CQLLibraryExportDAO.java
#	src/mat/dao/impl/clause/CQLLibraryExportDAOImpl.java
#	src/mat/server/bonnie/BonnieServiceImpl.java


Former-commit-id: 86661fcd6fb43cef6650ae07382680774554c101

+++ src/mat/server/bonnie/BonnieServiceImpl.java
@@ -1,25 +1,41 @@
 package mat.server.bonnie;
 
 import java.io.IOException;
+import java.util.Date;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import mat.client.bonnie.BonnieService;
+import mat.client.umls.service.VsacTicketInformation;
 import mat.dao.UserBonnieAccessInfoDAO;
 import mat.dao.UserDAO;
+import mat.dao.clause.MeasureDAO;
 import mat.model.User;
 import mat.model.UserBonnieAccessInfo;
+import mat.model.clause.Measure;
 import mat.server.LoggedInUserUtil;
 import mat.server.SpringRemoteServiceServlet;
+import mat.server.VSACAPIServiceImpl;
 import mat.server.bonnie.api.BonnieAPI;
 import mat.server.bonnie.api.BonnieAPIv1;
+import mat.server.bonnie.api.result.BonnieCalculatedResult;
+import mat.server.bonnie.api.result.BonnieMeasureResult;
+import mat.server.export.ExportResult;
 import mat.server.service.EncryptDecryptToken;
+import mat.server.service.SimpleEMeasureService;
 import mat.shared.BonnieOAuthResult;
+import mat.shared.bonnie.error.BonnieAlreadyExistsException;
+import mat.shared.bonnie.error.BonnieBadParameterException;
+import mat.shared.bonnie.error.BonnieDoesNotExistException;
+import mat.shared.bonnie.error.BonnieNotFoundException;
 import mat.shared.bonnie.error.BonnieServerException;
 import mat.shared.bonnie.error.BonnieUnauthorizedException;
+import mat.shared.bonnie.error.UMLSNotActiveException;
 import mat.shared.bonnie.result.BonnieUserInformationResult;
 
 @SuppressWarnings("serial")
@@ -40,6 +56,16 @@ public class BonnieServiceImpl extends SpringRemoteServiceServlet implements Bon
 
 	@Autowired
 	private UserBonnieAccessInfoDAO userBonnieAccessInfoDAO;
+	
+	@Autowired
+	private MeasureDAO measureDAO;
+	
+	@Autowired
+	private VSACAPIServiceImpl vasacAPI;
+	
+	@Autowired
+	@Qualifier("simpleEMeasureServiceImpl")
+	private SimpleEMeasureService simpleEMeasureService;
 
 	@Override
 	public String getBonnieAccessLink() {
@@ -58,7 +84,6 @@ public class BonnieServiceImpl extends SpringRemoteServiceServlet implements Bon
 
 	@Override
 	public BonnieOAuthResult exchangeCodeForTokens(String code) {
-		System.out.println("EXCHANING!");
 		BonnieOAuthResult result = null;
 		try {
 			result = bonnieApi.getBonnieOAuthResult(code);
@@ -71,7 +96,86 @@ public class BonnieServiceImpl extends SpringRemoteServiceServlet implements Bon
 
 		return result;
 	}
-
+	
+	private UserBonnieAccessInfo getUserBonnieAccessInfo(String userId) throws BonnieUnauthorizedException {
+		UserBonnieAccessInfo userInformation = null;
+		try {
+			userInformation = validateOrRefreshBonnieTokensForUser(userId);
+		} catch (BonnieUnauthorizedException e) {
+			handleBonnieUnauthorizedException(userInformation);
+			throw e;
+		}
+		return userInformation;
+	}
+	
+	@Override
+	public String getUpdateOrUploadMeasureToBonnie(String measureId, String userId, VsacTicketInformation vsacTicket) throws BonnieUnauthorizedException, BonnieBadParameterException, BonnieDoesNotExistException, BonnieServerException, IOException, BonnieAlreadyExistsException, UMLSNotActiveException {
+		
+		String successMessage = "";
+		
+		//see if user valid
+		UserBonnieAccessInfo userInformation = getUserBonnieAccessInfo(userId);
+		String userAccessToken = userInformation.getAccessToken();
+		
+		//Get measure and check if it is in bonnie
+		Measure measure = measureDAO.find(measureId);
+		String measureSetId = measure.getMeasureSet().getId();
+		BonnieMeasureResult bonnieMeasureResults = null;
+		try {
+			bonnieMeasureResults = bonnieApi.getMeasureById(userAccessToken, measureSetId);
+		} catch (BonnieUnauthorizedException e) {
+			handleBonnieUnauthorizedException(userInformation);
+			throw e;
+		}
+		
+		Date exportDate = new Date();
+		ExportResult export = null;
+		try {
+			export = simpleEMeasureService.getEMeasureZIP(measureId, exportDate);
+		} catch (Exception e1) {
+			throw new BonnieServerException();
+		}
+		
+		String currentReleaseVersion = StringUtils.replace(measure.getReleaseVersion(), ".", "_");
+		
+		byte[] zipFileContents = export.zipbarr;
+		String fileName = export.measureName + "_" + currentReleaseVersion + ".zip";
+		String calculationType = measure.getPatientBased() ? "patient" : "episode";
+		String vsacTicketGrantingTicket = vsacTicket.getTicket();
+		String vsacTicketExpiration = String.valueOf(vsacTicket.getTimeout().getTime());
+		
+		
+		if(bonnieMeasureResults != null && bonnieMeasureResults.getMeasureExsists()) {
+			//Measure exists in Bonnie Update
+			bonnieApi.updateMeasureInBonnie(userAccessToken, measureSetId, zipFileContents, fileName, null, calculationType, vsacTicketGrantingTicket, vsacTicketExpiration);
+			successMessage = measure.getDescription() +  " has been successfully updated in Bonnie. Please select open or save to view the results.";
+		} else {
+			//Measure doesn't exist in Bonnie upload measure
+			bonnieApi.uploadMeasureToBonnie(userAccessToken, zipFileContents, fileName, null, calculationType, vsacTicketGrantingTicket, vsacTicketExpiration);
+			successMessage = measure.getDescription() + " has been successfully uploaded as a new measure in Bonnie. Please go to the Bonnie tool to create test cases for this measure.";
+		}
+		
+		return successMessage;
+		
+	}
+	
+	public BonnieCalculatedResult getBonnieExportForMeasure(String userId, String measureId) throws IOException, BonnieUnauthorizedException, BonnieNotFoundException, BonnieServerException{
+		UserBonnieAccessInfo userInformation = getUserBonnieAccessInfo(userId);
+		String userAccessToken = userInformation.getAccessToken();
+		BonnieCalculatedResult caluclatedResult = null;
+		try {
+			caluclatedResult = bonnieApi.getCalculatedResultsForMeasure(userAccessToken, measureId);
+		} catch (BonnieUnauthorizedException e) {
+			handleBonnieUnauthorizedException(userInformation);
+			throw e;
+		} catch (BonnieNotFoundException e) {
+			throw e;
+		} catch (BonnieServerException e) {
+			throw e;
+		}
+		return caluclatedResult;
+	}
+	
 	private BonnieOAuthResult refreshBonnieTokens(String userId) throws BonnieUnauthorizedException {
 		BonnieOAuthResult result = null;
 		User user = userDAO.find(userId);
@@ -126,7 +230,9 @@ public class BonnieServiceImpl extends SpringRemoteServiceServlet implements Bon
 		// if an unauthorized exception is thrown and the user had credentials in the
 		// database, delete them because they
 		// are invalid, and the surface the error
-		userBonnieAccessInfoDAO.delete(Integer.toString(bonnieAccessInfo.getId()));
+		if(bonnieAccessInfo != null) {
+			userBonnieAccessInfoDAO.delete(Integer.toString(bonnieAccessInfo.getId()));
+		}
 	}
 
 	public BonnieAPI getBonnieApi() {
